+++ smiianProj/src/main/java/pages/PPLoginPage.java
@@ -32,28 +32,21 @@ public class PPLoginPage extends PPParentPage {
             loggerExam.info("PrivatPage is opened");
         } catch (Exception e) {
             loggerExam.error("Can't open Privat Page" + e);
-//            Assert.fail("Can't open Privat Page" + e);           //  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
         }
         return this;
     }
 
-//    public WebElement getCurrencyValueONE (String CurrencyText){
-//        return webDriver.findElement(By.xpath(String.format(CurrencyText, currency)));
-//    }
 
     public void getUiCurrencyBuyValueAndSave(String currencyText){            //  .//td[@id='USD_buy']
         WebElement element = webDriver.findElement(By.xpath(".//td[@id='" + currencyText + "_buy']"));
 //        int currencyBuyValue = Integer.parseInt(element.getText());
-        TestData.uiCurrencyValueBuy = Integer.parseInt(element.getText());
+        TestData.uiCurrencyValueBuy = Double.parseDouble(element.getText());
 //        return Integer.parseInt(element.getText());
     }
 
     public void getUiCurrencySellValueAndSave(String currencyText){            //  .//td[@id='USD_sell']
         WebElement element = webDriver.findElement(By.xpath(".//td[@id='" + currencyText + "_sell']"));
-        TestData.uiCurrencyValueSell = Integer.parseInt(element.getText());
+        TestData.uiCurrencyValueSell = Double.parseDouble(element.getText());
     }
-
-
-
 
 }
